package: supply runelite and lombok as pre-packaged jars
This opts them out of verification and resolution by the plugin's build
so we can safely bump any part of them without breaking the release
builds of the plugins. This also lets us build the hub locally with a
local version of RuneLite to test for api breakage, as some plugins
have mavenLocal disabled.